february-2021-leetcode-challenge/src/test/java/org/phoenix/leetcode/challenges/Problem28_MaximumFrequencyStackTest.java
package org.phoenix.leetcode.challenges;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.phoenix.leetcode.challenges.Problem28_MaximumFrequencyStack.FreqStack;

class Problem28_MaximumFrequencyStackTest {

    private final FreqStack stack = new FreqStack();

    @Test
    void test() {
        stack.push(5);
        stack.push(7);
        stack.push(5);
        stack.push(7);
        stack.push(4);
        stack.push(5);
        assertEquals(5, stack.pop());
        assertEquals(7, stack.pop());
        assertEquals(5, stack.pop());
        assertEquals(4, stack.pop());
    }
}